Fix creation of libsystemc.a with llvm-link.
The script used to "add" modules to the existing libsystemc.a in a loop,
thus writting and reading to the same file with llvm-link. This seemed to
create an incomplete archive. We now call llvm-link just once on all the
*.a files.

While we're at it, we also simplify the installation and allow
$INSTALL_PATH_SYSTEMC_LLVM to be the same path as the source dir.