bitcode loader: don't try to load libstdc++.
The problem we tried to solve was to be able to run elaboration calling
libstdc++ functions from the JIT, which was simply buggy in LLVM 2.6. The
new version solves this issue, and we're now able to run an elaboration
calling "cout << "bla" << endl;".

The bitcode loader still crashes, but does so after the callback to the
front-end.