Compile PinaVM as a dynamic library + dummy executable.
The platform needs to access the SystemC library contained in the
executable. The way to do that is through dlopen(NULL, ...), wrapped in
DynamicLibrary::LoadLibraryPermanently(NULL, ...), but this doesn't show
the executable's symbols, just the library it links against. Putting all
the stuff inside a dynamic library allows showing everything to the
JIT-ed code. In particular, pinapa_callback is now directly visible, we
don't need to DynamicLibrary::AddSymbol() it anymore.

This allows one to load just the platform's bitcode, which is able to
access the SystemC library. However, the front-end's JIT management
requires the bitcode for SystemC (to know about the SystemC types), so
right now, the code crashes in the JIT stuff of the front-end.