New -scbc option to load SystemC bitcode.
The elaboration running code needs to have _only_ the platform's bitcode,
to make sure the accesses to the SystemC library (registering new
sc_objects in particual) do access the symbols of the SystemC library
loaded in PinaVM.

OTOH, the JIT stuff in the front-end needs to have access to the bitcode
for the whole SystemC+platform, to have the typing information (most
JIT-ed functions return SystemC objects, and need to be able to talk
about them ...).

For now, we load both bitcode files (the new -scbc FILE.bc takes the
linked version, and the positional argument takes just the platform). In
the long run, we should load the platform's bitcode first, and load
libsystemc.a (bitcode version) afterwards, in the same execution
engine (we have two separate execution engines at the moment).

The platform jerome-chain's Makefile is modified accordingly, and "make
promela" now WORKS with the bitcode loader.

YIPEEE!!!