example Makefile: Don't let targets depend on $(PINAVM)
The idea of this dependancy was to make sure pinavm was up-to-date before
running it, but since we don't have the set of source files available
here, we just consider pinavm to be a .PHONY target, which implies that
it's not considered up-to-date even when there's nothing to do.

Instead, we add a recursive -$(MAKE) --no-print-directory $(PINAVM)
before running pinavm. It's almost instantaneous, and doesn't trigger
useless rebuilds.

The fix is fully applied only to events/.