example Makefile: run ulimit-ed commands from a subshell.
It's not a good idea to run ulimit within a loop: the limit applies to
the current shell and its child processes. Using a subshell (the (...)
notation), the ulimit is applied to each run of the loop independantly.
It also allows to get rid of the "cd ..", and should allow variable-depth
subdirectories (which would have required cd ../.. or so).