Fix data-width bug in jitBool.
jitBool creates a function that returns a i8 type. In practice, it will
most likely return this into a 32-bits register (%eax on i386) with high
bits unset. If we get the value into an int, we get arbitrary value for
these high bits, and the result is broken.

I guess the way to return the result (essentially a "return !res;") was
an attempt to fix the bug, it's no longer necessary (otherwise, we get
the exact opposite of what we want).