An example platform using an external library.
Calls to external library functions in process body are intrinsequely
hard to manage, but library function calls in elaboration are not a
problem: pinavm already has a -load option (thanks to
library before loading the bytecode. Then, the LLVM JIT engine will
resolve symbols in the bitcode using this library.

We illustrate this with an example using a dummy library. To keep the
Makefile simple, we add a PINAVM_LIBS variable and use it in common.mk.