Allow separate compilation in common.mk
We keep the old behavior is SRC is defined but not SRCS. If SRCS is
defined, then we generate one .bc file per .cpp, and link the result with
llvm-link. We chose the name of the .bc file that will be used as input
to PinaVM in $(PINAVM_INPUT_BC).