Use CallSite::getArgument, not CallInst::getOperand.
CallInst::getOperand gives operand number N looking at the call as an
LLVM construct. The order changed in LLVM 2.8, which essentially means
that any call to CallInst::getOperand written for LLVM 2.7 is broken now.

In particular, in the case of ASSERT(false), "callInst->getOperand(1)"
used to return a pointer to "false", and now returns a pointer to ASSERT,
seen as a pointer to function.

PinaVM was broken even on demo_opentlm before this commit, but
demo_opentlm now works again with the SPIN back-end.