An example where dynamic_cast between JIT-ed code and native code
I'm getting the same problem in PinaVM itself, but reproducing it here
for clarity:

dynamic_cast needs access to the vtable. When dynamic_cast is ran from
the native code, it seems there are cases when it cannot access the
information which is available on the bitcode side (because the same
dynamic_cast succeeds in JIT mode).

Based on this, the problem occurs only when dynamic_cast-ing A to B
with :

* A has been created within the JIT engine.

* Either the type of A or the one of B has no virtual function defined in
  the .cpp file

The vtable symbols are defined as 'W' in the bitcode. man llvm-nm says
this:

       W   Weak reference (multiple definitions link together into zero or one definitions)

My understanding is that at load time, the weak references should be
linked against references in the executable, but it is not the case.