BASIC: don't claim the bus is bound to a channel
The BASIC bus is technically an sc_module, but is considered as an
"interface" in the IR of PinaVM. Before this change, the ports of the bus
were considered to be bound to an uninitialized pointer (=> segfault).
For now, just let a NULL pointer as the channel associated to these ports.