Port the build system to CMake
The new system hasn't been tested outside Linux, but should be a first
step towards something more portable, with better integration in IDEs.

We also get the CMake goodies, like dependancies management almost for
free, progress bars, colors, ... and a relatively user-friendly syntax
compared to the previous Makefile.

We still rely on SystemC's build system to compile SystemC itself. Making
a CMake-based build script for SystemC would be simple, but is not
totally immediate because of the pieces of assembly it contains.