Modify backends to use ALLConstruct
After moving all the constructs to SCConstructs/, we need a easy way
to include all the constructs and handlers. One day when we add a new
construct, we don't manually add new headers to each backends. Just
add it to frontend/ALLConstruct.h

ALLConstruct is a file which includes all constructs.
ALLHandler is a file which includes all handlers.