Write document about how to add new backend
Adding a new backend may be a complex work before. I think it is
necessary to document the way to add new backend after commit d661dfe.

I also think it is good to show a hello world example, so I implement
HelloBackend, which gives users a tangible example for how to create
a new backend.