Port SimpleBackendNameAllUsedStructsAndMergeFunctions to LLVM 3.2
Originally, this class try to remove type name from TypeSymbolTable.
This way is impossible in LLVM 3.2, because all the names of StructType
is keeped by LLVMContextImpl, LLVMContextImpl is hidden from API so
that we can not access it directly.

Luckily, set the StructType name with zero size string can delete
the StructType name from Table of LLVMContextImpl.

Corresponding part of llvm source:
http://llvm.org/docs/doxygen/html/Type_8cpp_source.html#l00441