Rearrange Code for better readability
I find there are some bugs on FunctionBuilder.
If PinaVM parse a syntax like :

int a() {return 1;}
int b() {return 1;}

Module::Process() {
	Port.write( a() && b() );
}

It will fail due to the PHINode of JIT_Function is broken. The
incomming BasicBlocks is less than the BasicBlocks PHINode specified.
I just Rearrange Code to find out where the bugs.  This commit should
not change the behavior of PinaVM.