Modify installLLVM.cmake to compatible with LLVM 3.3
Clang has been renamed to cfe(C Frontend) in LLVM 3.3 release. I
must modify the autoinstall script to make cmake smartly choose
the right directory to extract tgz file.

Originally, if LLVM version is different from the recommended
version, it just shows a warning. Now, I think it is better to
show a fatal error or autoinstall the right LLVM for you. Providing
autoinstall function if wrong version means we may call autoinstall
twice, one for finding no llvm-config, the other for wrong LLVM
version. It give me a reason to warp installLLVM.cmake as a
CMake function, so that installLLVM.cmake is not included twice.